requirements.txt
numpy==1.18.1
transformers==2.5.1
python-igraph==0.7.1.post6
pytorch-transformers==1.2.0
networkx==2.4.0
scipy==1.2.3